# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Faydamed API Authentication",
        "description": "Authentication API endpoints for Faydamed system",
        "contact": {
            "email": "support@faydamed.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        },
        {
            "url": "https://faydaapi.dotprogrammers.com",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/admin/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get all blogs",
                "description": "Retrieve a list of all blogs (Admin only)",
                "operationId": "getAllBlogs",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "published",
                                "schedule"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blogs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blogs retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Blog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create a new blog",
                "description": "Create a new blog post (Admin only)",
                "operationId": "createBlog",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "category",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Introduction to Laravel"
                                    },
                                    "url_slug": {
                                        "description": "Auto-generated from title if not provided",
                                        "type": "string",
                                        "example": "introduction-to-laravel"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Technology"
                                    },
                                    "excerpt": {
                                        "type": "string",
                                        "example": "A brief introduction to Laravel framework"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Laravel is a web application framework..."
                                    },
                                    "tags": {
                                        "description": "JSON string array of tags",
                                        "type": "string",
                                        "example": "[\"laravel\",\"php\",\"framework\"]"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "schedule"
                                        ],
                                        "example": "draft"
                                    },
                                    "published_at": {
                                        "description": "Required for schedule status",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-10T10:00:00Z"
                                    },
                                    "blog_images[]": {
                                        "description": "Multiple blog images",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Blog created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Blog created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Blog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Blog": {
                "title": "Blog",
                "description": "Blog post model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Introduction to Laravel"
                    },
                    "url_slug": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "introduction-to-laravel-1699456789"
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "category": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Technology"
                    },
                    "excerpt": {
                        "type": "string",
                        "example": "A brief introduction to Laravel framework and its features",
                        "nullable": true
                    },
                    "content": {
                        "type": "string",
                        "example": "Laravel is a powerful PHP framework that makes web development easier..."
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "laravel",
                            "php",
                            "framework",
                            "web development"
                        ],
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "schedule"
                        ],
                        "example": "published"
                    },
                    "blog_images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "blogs/images/1699456789_abc123.jpg",
                            "blogs/images/1699456790_def456.jpg"
                        ],
                        "nullable": true
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-08T10:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-08T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-08T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}